+++ reus/fbref/fb_player_match_logs_passing_type.py
@@ -2,9 +2,20 @@ from ..util import get_page_soup
 from .util import match_log_iterator
 
 
-def fb_player_match_logs_passing_types(
+def fb_player_match_logs_passing_type(
     pageSoup=None, season_end: str = None, player_id: str = None
 ) -> list:
+    """Retrieves a players passing type match log for a given season
+
+    Args:
+        pageSoup (bs4, optional): bs4 object of a players passing type match log page. Defaults to None.
+        season_end (str): ending year of a season
+        player_id (str): unique identifier for a player
+
+    Returns:
+        list: passing type match log for a player in a given season
+    """
+
     url = f"https://fbref.com/en/players/{player_id}/matchlogs/{int(season_end)-1}-{season_end}/passing_types/"
 
     assert (
@@ -51,6 +62,3 @@ def fb_player_match_logs_passing_types(
     mylist = match_log_iterator(rows=rows, attributes=attributes)
 
     return mylist
-
-
-# TODO: documentation