+++ reus/fbref/fb_player_match_logs_summary.py
@@ -5,6 +5,17 @@ from .util import match_log_iterator
 def fb_player_match_logs_summary(
     pageSoup=None, season_end: str = None, player_id: str = None
 ) -> list:
+    """Retrieves a players summary match log for a given season
+
+    Args:
+        pageSoup (bs4, optional): bs4 object of a players summary log page. Defaults to None.
+        season_end (str): ending year of a season
+        player_id (str): unique identifier for a player
+
+    Returns:
+        list: summary log for a player in a given season
+    """
+
     url = f"https://fbref.com/en/players/{player_id}/matchlogs/{int(season_end)-1}-{season_end}/"
 
     assert (
@@ -66,5 +77,3 @@ if __name__ == "__main__":
     import pandas as pd
 
     print(pd.DataFrame(data))
-
-# TODO: documentation