+++ reus/transfermarkt/__init__.py
@@ -4,4 +4,6 @@ from .tm_player_market_value import tm_player_market_value  # noqa: F401
 from .tm_player_metadata import tm_player_metadata  # noqa: F401
 from .tm_player_transfers import tm_player_transfers  # noqa: F401
 from .tm_team_player_data import tm_team_player_data  # noqa: F401
+from .tm_team_staff import tm_team_staff  # noqa: F401
+from .tm_team_staff_history import tm_team_staff_history  # noqa: F401
 from .tm_team_transfers import tm_team_transfers  # noqa: F401