+++ tests/fotmob/test_fm_league_matches.py
@@ -1,9 +1,11 @@
-import unittest
-import numpy.testing as npt
-from pathlib import Path
-import pandas as pd
-from reus.fotmob import fm_league_matches
 import time
+import unittest
+from pathlib import Path
+
+import numpy.testing as npt
+import pandas as pd
+
+from reus.fotmob import fm_league_matches
 
 
 class TestFmLeagueMatches(unittest.TestCase):
@@ -14,7 +16,7 @@ class TestFmLeagueMatches(unittest.TestCase):
         expected = pd.read_csv(self.data_path / "epl_matches_2022_2023.csv")
         actual = fm_league_matches(league_id=47, season="2022/2023")
 
-        drop_cols = ["round", "roundName", "utcTime"]
+        drop_cols = ["pageUrl", "round", "roundName", "utcTime"]
         expected.drop(columns=drop_cols, inplace=True)
         actual.drop(columns=drop_cols, inplace=True)
 
@@ -28,7 +30,7 @@ class TestFmLeagueMatches(unittest.TestCase):
         expected = pd.read_csv(self.data_path / "mls_matches_2022.csv")
         actual = fm_league_matches(league_id=130, season="2022")
 
-        drop_cols = ["round", "roundName", "utcTime"]
+        drop_cols = ["pageUrl", "round", "roundName", "utcTime"]
         expected.drop(columns=drop_cols, inplace=True)
         actual.drop(columns=drop_cols, inplace=True)