+++ tests/fotmob/test_fm_season_stats.py
@@ -1,26 +1,28 @@
-import unittest
-import numpy.testing as npt
-from pathlib import Path
 import json
-from reus.fotmob import fm_season_stats
 import time
+import unittest
+from pathlib import Path
+
+import numpy.testing as npt
+
+from reus.fotmob import fm_season_stats
 
 
-class TestFmSeasonStatLeaders(unittest.TestCase):
+class TestFmSeasonStats(unittest.TestCase):
     current_dir = Path(__file__).resolve().parent
     data_path = current_dir / "data"
 
-    def test_epl_big_chances_created_leaders(self):
+    def test_epl_big_chances_created(self):
         with open(
             self.data_path
-            / "epl_big_chances_created_complete_season_stat_2020_2021.json"
+            / "epl_big_chances_created_complete_season_stat_2020_2021.json"  # noqa: W503
         ) as f:
             expected = json.load(f)
 
         actual = fm_season_stats(
             league_id=47,
             team_or_player="teams",
-            stat_name="Big chances created",
+            stat_name="Big chances",
             season="2020/2021",
         ).to_dict(orient="records")